pyproject.toml
[tool.poetry]
name = "historynerf"
version = "0.1.0"
description = ""
authors = ["Bradley-Butcher <bradleydpbutcher@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pycolmap = "^0.4.0"
hydra-core = "^1.3.2"
numpy = "^1.24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
